+++ .gitignore
@@ -28,3 +28,4 @@ hs_err_pid*
 gradle/*
 jav.iml
 gradlew*
+gradle-wrapper*
